game score 타입 변경

+++ src/main/kotlin/kr/jiasoft/hiteen/feature/play/domain/GameScoreEntity.kt
@@ -11,7 +11,7 @@ data class GameScoreEntity(
     val seasonId: Long,
     val participantId: Long,
     val gameId: Long,
-    val score: Long,
+    val score: Double,
     val tryCount: Int = 1,
     val createdAt: OffsetDateTime = OffsetDateTime.now(),
     val updatedAt: OffsetDateTime? = null,
